protecting routes with auth middleware

+++ backend/src/controllers/notesController.ts
@@ -2,6 +2,7 @@ import { RequestHandler } from "express";
 import Note from "../models/noteModel";
 import createHttpError from "http-errors";
 import mongoose from "mongoose";
+import { assertIsDefined } from "../utils/assertIsDefined";
 
 interface NoteBody {
   title?: string;
@@ -9,9 +10,11 @@ interface NoteBody {
 }
 
 export const getAllNotes: RequestHandler = async (req, res, next) => {
-  console.log(req);
+  const authenticatedUserId = req.session.userId;
+
   try {
-    const notes = await Note.find();
+    assertIsDefined(authenticatedUserId);
+    const notes = await Note.find({ createdBy: authenticatedUserId });
     res.status(200).json(notes);
   } catch (error) {
     next(error);
@@ -20,11 +23,17 @@ export const getAllNotes: RequestHandler = async (req, res, next) => {
 
 export const getNote: RequestHandler = async (req, res, next) => {
   try {
+    const authenticatedUserId = req.session.userId;
+
     if (!mongoose.isValidObjectId(req.params.noteId))
       throw createHttpError(400, "Invalid note id");
 
     const note = await Note.findById(req.params.noteId);
     if (!note) throw createHttpError(404, "Note not found");
+
+    if (!note.createdBy.equals(authenticatedUserId)) {
+      throw createHttpError(401, "You cant access this note");
+    }
 
     res.status(200).json(note);
   } catch (error) {
@@ -39,9 +48,16 @@ export const createNote: RequestHandler<
   unknown
 > = async (req, res, next) => {
   const { title, text } = req.body;
+  const authenticatedUserId = req.session.userId;
+
   try {
+    assertIsDefined(authenticatedUserId);
     if (!title) throw createHttpError(400, "Note must have a title");
-    const note = await Note.create({ title, text });
+    const note = await Note.create({
+      title,
+      text,
+      createdBy: req.session.userId,
+    });
     res.status(201).json(note);
   } catch (error) {
     next(error);
@@ -58,7 +74,10 @@ export const updateNote: RequestHandler<
   NoteBody,
   unknown
 > = async (req, res, next) => {
+  const authenticatedUserId = req.session.userId;
+
   try {
+    assertIsDefined(authenticatedUserId);
     if (!mongoose.isValidObjectId(req.params.noteId))
       throw createHttpError(400, "Invalid note id");
 
@@ -75,6 +94,9 @@ export const updateNote: RequestHandler<
 
     if (!updatedNote) throw createHttpError(404, "Note not found");
 
+    if (!updatedNote.createdBy.equals(authenticatedUserId)) {
+      throw createHttpError(401, "You cant access this note");
+    }
     res.status(200).json(updatedNote);
   } catch (error) {
     next(error);
@@ -82,12 +104,19 @@ export const updateNote: RequestHandler<
 };
 
 export const deleteNote: RequestHandler = async (req, res, next) => {
+  const authenticatedUserId = req.session.userId;
+
   try {
+    assertIsDefined(authenticatedUserId);
+
     if (!mongoose.isValidObjectId(req.params.noteId))
       throw createHttpError(400, "Invalid note id");
 
     const note = await Note.findById(req.params.noteId);
     if (!note) throw createHttpError(404, "Note not found");
+    if (!note.createdBy.equals(authenticatedUserId)) {
+      throw createHttpError(401, "You cant access this note");
+    }
 
     await note.deleteOne();
